+++ src/main/java/onlyoffice/OnlyOfficeEditorServlet.java
@@ -107,6 +107,7 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
 
         properties = configurationManager.getProperties();
 
+        String type = "";
         String callbackUrl = "";
         String fileUrl = "";
         String gobackUrl = "";
@@ -147,6 +148,8 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
             user = AuthenticatedUserThreadLocal.get();
             log.info("user " + user);
             if (attachmentUtil.checkAccess(attachmentId, user, false)) {
+                type = documentManager.getEditorType(request.getHeader("USER-AGENT"));
+
                 key = documentManager.getKeyOfFile(attachmentId);
 
                 fileName = attachmentUtil.getFileName(attachmentId);
@@ -174,10 +177,10 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
         response.setContentType("text/html;charset=UTF-8");
         PrintWriter writer = response.getWriter();
 
-        writer.write(getTemplate(apiUrl, callbackUrl, fileUrl, key, fileName, user, gobackUrl, errorMessage));
+        writer.write(getTemplate(type, apiUrl, callbackUrl, fileUrl, key, fileName, user, gobackUrl, errorMessage));
     }
 
-    private String getTemplate(String apiUrl, String callbackUrl, String fileUrl, String key, String fileName,
+    private String getTemplate(String type, String apiUrl, String callbackUrl, String fileUrl, String key, String fileName,
             ConfluenceUser user, String gobackUrl, String errorMessage) throws UnsupportedEncodingException {
         Map<String, Object> defaults = MacroUtils.defaultVelocityContext();
         Map<String, String> config = new HashMap<String, String>();
@@ -199,7 +202,7 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
         JSONObject gobackObject = new JSONObject();
 
         try {
-            responseJson.put("type", "desktop");
+            responseJson.put("type", type);
             responseJson.put("width", "100%");
             responseJson.put("height", "100%");
             responseJson.put("documentType", documentManager.getDocType(docExt));